+++ tests/no_compile/inner_mmio_double_borrow.rs
@@ -11,9 +11,9 @@ struct UartBank {
 struct Uart {
     // you can be explicit if you like
     control: u32,
-    #[mmio(inner)]
+    #[mmio(Inner)]
     bank_0: UartBank,
-    #[mmio(inner)]
+    #[mmio(Inner)]
     bank_1: UartBank,
 }